+++ src/DevShop/Component/PowerProcess/composer.json
@@ -26,7 +26,7 @@
     },
     "autoload": {
         "psr-4": {
-            "ProvisionOps\\Tools\\": "src/"
+            "DevShop\\Component\\PowerProcess\\": ""
         }
     },
     "config": {
